requirements.txt
altgraph==0.17.4
base==0.0.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pefile==2023.2.7
pillow==11.2.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
pyparsing==3.2.3
PyQt6==6.9.0
PyQt6-Qt6==6.9.0
PyQt6_sip==13.10.0
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
requests==2.32.3
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
sympy==1.14.0
thop==0.1.1.post2209072238
torch==2.7.0
torchaudio==2.7.0+cu118
torchvision==0.22.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.128
ultralytics-thop==2.0.14
urllib3==2.4.0
zipp==3.21.0
